Use autoBind, pass context rather than manually binding
