+++ test/unit/step.spec.js
@@ -132,164 +132,6 @@ describe('Tour | Step', () => {
     });
   });
 
-  describe('bindAdvance()', () => {
-    let event;
-    let link;
-    let hasAdvanced = false;
-
-    const advanceOnSelector = 'test-selector';
-    const advanceOnEventName = 'test-event';
-    const tourProto = {
-      next() { hasAdvanced = true; }
-    };
-
-    beforeEach(() => {
-      event = new Event(advanceOnEventName);
-
-      link = document.createElement('a');
-      link.classList.add(advanceOnSelector);
-      link.textContent = 'Click Me 👋';
-
-      document.body.appendChild(link);
-    });
-
-    afterEach(() => {
-      link.remove();
-    });
-
-    it('triggers the `advanceOn` option via object', () => {
-      const step = new Step(tourProto, {
-        advanceOn: { selector: `.${advanceOnSelector}`, event: advanceOnEventName }
-      });
-
-      step.isOpen = () => true;
-
-      step.bindAdvance();
-      link.dispatchEvent(event);
-
-      expect(link.classList.contains(advanceOnSelector)).toBe(true);
-      expect(hasAdvanced, '`next()` triggered for advanceOn').toBe(true);
-    });
-
-    it('captures events attached to no element', () => {
-      const step = new Step(tourProto, {
-        advanceOn: { event: advanceOnEventName }
-      });
-
-      step.isOpen = () => true;
-
-      step.bindAdvance();
-      document.body.dispatchEvent(event);
-
-      expect(hasAdvanced, '`next()` triggered for advanceOn').toBeTruthy();
-    });
-
-    it('should support bubbling events for nodes that do not exist yet', () => {
-      const event = new Event('blur');
-
-      const step = new Step(tourProto, {
-        text: 'Lorem ipsum dolor: <a href="https://example.com">sit amet</a>',
-        advanceOn: {
-          selector: 'a[href="https://example.com"]',
-          event: 'blur'
-        }
-      });
-
-      step.isOpen = () => true;
-
-      step.bindAdvance();
-      document.body.dispatchEvent(event);
-
-      expect(hasAdvanced, '`next()` triggered for advanceOn').toBeTruthy();
-    });
-
-    it('calls `removeEventListener` when destroyed', function(done) {
-      const bodySpy = spy(document.body, 'removeEventListener');
-      const step = new Step(tourProto, {
-        advanceOn: { event: advanceOnEventName }
-      });
-
-      step.isOpen = () => true;
-
-      step.bindAdvance();
-      step.trigger('destroy');
-
-      expect(bodySpy.called).toBe(true);
-      bodySpy.restore();
-
-      done();
-    });
-  });
-
-  describe('bindButtonEvents()', () => {
-    const link = document.createElement('a');
-    const step = new Step(new Tour(), {});
-    it('adds button events', () => {
-      const event = new Event('test');
-      const hover = new Event('mouseover');
-      let eventTriggered = false;
-
-      step.bindButtonEvents({
-        events: {
-          'mouseover': '1',
-          test: () => eventTriggered = true
-        },
-        text: 'Next',
-        action: () => {}
-      }, link);
-
-      link.dispatchEvent(event);
-      link.dispatchEvent(hover);
-      expect(eventTriggered, 'custom button event was bound/triggered').toBeTruthy();
-    });
-
-    it('removes events once destroyed', () => {
-      step.destroy();
-
-      expect(link.hasAttribute('data-button-event'), 'attribute to confirm event is removed').toBeFalsy();
-    });
-
-  });
-
-  describe('bindCancelLink()', () => {
-    it('adds an event handler for the cancel button', () => {
-      const event = new MouseEvent('click', {
-        view: window,
-        bubbles: true,
-        cancelable: true
-      });
-      const link = document.createElement('a');
-      const step = new Step();
-      let cancelCalled = false;
-
-      step.cancel = () => cancelCalled = true;
-      step.bindCancelLink(link);
-
-      link.dispatchEvent(event);
-      expect(cancelCalled, 'cancel method was called from bound click event').toBeTruthy();
-    });
-  });
-
-  describe('bindMethods()', () => {
-    it('binds the expected methods', () => {
-      const step = new Step();
-      const methods = [
-        '_scrollTo',
-        '_setupElements',
-        '_show',
-        'cancel',
-        'complete',
-        'destroy',
-        'hide',
-        'isOpen',
-        'show'
-      ];
-      methods.forEach((method) => {
-        expect(step[method], `${method} has been bound`).toBeTruthy();
-      });
-    });
-  });
-
   describe('cancel()', () => {
     it('triggers the cancel event and tour method', () => {
       let cancelCalled = false;
@@ -358,17 +200,6 @@ describe('Tour | Step', () => {
     });
   });
 
-  describe('parseAttachTo()', function() {
-    it('fails if element does not exist', function() {
-      const step = new Step({}, {
-        attachTo: { element: '.scroll-test', on: 'center' }
-      });
-
-      const { element } = step.parseAttachTo();
-      expect(element).toBeFalsy();
-    });
-  });
-
   describe('_setupElements()', () => {
     it('calls destroy on the step if the content element is already set', () => {
       const step = new Step();
@@ -387,18 +218,6 @@ describe('Tour | Step', () => {
       };
       step._setupElements();
       expect(destroyCalled, '_setupElements method called destroy on the existing tooltip').toBe(true);
-    });
-
-    it('calls bindAdvance() if advanceOn passed', () => {
-      const step = new Step({
-        next: () => true
-      }, {
-        advanceOn: { selector: '.click-test', event: 'test' }
-      });
-      const bindFunction = spy(step, 'bindAdvance');
-      step._setupElements();
-
-      expect(bindFunction.called).toBeTruthy();
     });
   });
 
@@ -581,13 +400,9 @@ describe('Tour | Step', () => {
       const element = document.createElement('div');
       const step = new Step(null, { showCancelLink: true });
 
-      const cancelLinkStub = stub(step, 'bindCancelLink');
-
       step._addCancelLink(element, header);
 
-      expect(cancelLinkStub.called).toBe(true);
       expect(element).toHaveClass('shepherd-has-cancel-link');
-      cancelLinkStub.restore();
     });
   });
